Se termina la implementación de creacion y login de usuarios usando parse [hay que implementar para testear]

+++ java/shrine/app/src/main/java/com/google/codelabs/mdc/java/shrine/Model/User.java
@@ -6,6 +6,6 @@ public class User {
     public String name;
     public String surname;
     public String email;
-    public float score;
+    public double score;
     public Bitmap image;
 }
